Operators and Decision Processing

+++ ObviousCode.Alchemy.Creatures/ObviousCode.Alchemy.Creatures.DecisionProcessing/Value/PredicateValue.cs
@@ -8,8 +8,8 @@ namespace ObviousCode.Alchemy.Creatures.DecisionProcessing
 		{
 			
 		}
-			
-		public abstract double GetValue ();
+
+		public abstract decimal GetValue ();
 	}
 }
